requirements.txt
# In-Car Multi-Speaker ASR System
# Python 3.10 | CUDA 11.8+
# Generated: 2026-04-23

# Core Deep Learning
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0

# HuggingFace ecosystem
transformers==4.36.2
huggingface-hub==0.20.3
datasets==2.16.1
accelerate==0.26.1

# Speech processing (ConvTasNet / Asteroid: optional — see requirements-extras.txt; avoids pesq build on some macOS setups)
speechbrain==0.5.16

# Audio I/O & processing
librosa==0.10.1
soundfile==0.12.1
audioread==3.0.1
resampy==0.4.2

# Evaluation metrics (CER/cpWER in src; scipy for Hungarian / cpWER)
jiwer==3.0.3
scipy==1.11.4

# Scientific computing
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.1

# Web UI
streamlit==1.30.0
streamlit-webrtc==0.47.1

# Configuration
omegaconf==2.3.0
pyyaml==6.0.1

# Code quality
flake8==7.0.0
mypy==1.8.0
black==24.1.1
isort==5.13.2

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
click==8.1.7
loguru==0.7.2
python-dotenv==1.0.0

# Optional: FastAPI backend
fastapi==0.109.0
uvicorn==0.27.0
